+++ src/image-generator.ts
@@ -407,7 +407,6 @@ export class ImageGenerator {
         const setDefinitions: { [modres: string]: SubstructureDef } = {};
         for (const modres in modresInfo) setDefinitions[modres] = modresInfo[modres].instances;
 
-        const unprocessedModres = new Set(Object.keys(modresInfo));
         const colorsIterator = cycleIterator(MODRES_COLORS);
         await using(structure.makeGroup({ label: 'Modified Residues' }), async group => {
             const modresStructures = await group.makeSubstructures(setDefinitions);
@@ -417,14 +416,14 @@ export class ImageGenerator {
                 struct.setCollapsed(ALLOW_COLLAPSED_NODES);
                 struct.setVisible(false);
             }
-            for (const [modres, struct] of Object.entries(modresStructures)) {
-                const nInstances = struct.data!.atomicResidueCount; // Using number of instances in the assembly, not in the deposited model
-                struct.setVisible(true);
+            for (const modres in modresInfo) {
+                const struct = modresStructures[modres] as StructureNode | undefined;
+                const nInstances = struct?.data?.atomicResidueCount ?? 0; // Using number of instances in the assembly, not in the deposited model
+                if (nInstances === 0) logger.warn(`Modified residue ${modres} is not present in the processed assembly (creating image without highlighted modified residue)`);
+                if (struct) struct.setVisible(true);
                 await this.saveViews('all', view => Captions.forModifiedResidue({ ...context, modresInfo: { ...modresInfo[modres], nInstances }, view }));
-                struct.setVisible(false);
-                unprocessedModres.delete(modres);
+                if (struct) struct.setVisible(false);
             };
-            if (unprocessedModres.size > 0) logger.error(`Failed to create images for these modified residues: ${Array.from(unprocessedModres).sort()}`);
         });
     }