+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed bug in checking expected files (do not check if empty when missing)
 - Do not fail when PDBe Structure Validation Report is not available, show gray structure instead
+- Do not fail when modified residue is not modelled, show structure without anything highlighted instead
 
 ## [2.1.0] - 2023-07-20